requirements.txt
anyio==4.9.0

asgiref==3.9.1

attrs==25.3.0

autobahn==24.4.2

Automat==25.4.16

CacheControl==0.14.3

cachetools==5.5.2

certifi==2025.7.14

cffi==1.17.1

channels==4.0.0

channels_auth_token_middlewares==1.2.0

channels_redis==4.2.1

charset-normalizer==3.4.2

constantly==23.10.4

cryptography==45.0.5

daphne==4.2.1

Django==5.2.4

django-cors-headers==4.7.0

djangorestframework==3.16.0

djangorestframework_simplejwt==5.5.0

firebase-admin==6.9.0

git-filter-repo==2.47.0

google-api-core==2.25.1

google-api-python-client==2.176.0

google-auth==2.40.3

google-auth-httplib2==0.2.0

google-cloud-core==2.4.3

google-cloud-firestore==2.21.0

google-cloud-storage==3.2.0

google-crc32c==1.7.1

google-resumable-media==2.7.2

googleapis-common-protos==1.70.0

grpcio==1.73.1

grpcio-status==1.73.1

h11==0.16.0

h2==4.2.0

hpack==4.1.0

httpcore==1.0.9

httplib2==0.22.0

httpx==0.28.1

hyperframe==6.1.0

hyperlink==21.0.0

idna==3.10

incremental==24.7.2

msgpack==1.1.1

pillow==11.3.0

proto-plus==1.26.1

protobuf==6.31.1

psycopg2-binary==2.9.10

pyasn1==0.6.1

pyasn1_modules==0.4.2

pycparser==2.22

PyJWT==2.9.0

pyOpenSSL==25.1.0

pyparsing==3.2.3

python-decouple==3.8

python-dotenv==1.1.1

razorpay==1.4.2

redis==6.2.0

requests==2.32.4

rsa==4.9.1

service-identity==24.2.0

setuptools==80.9.0

sniffio==1.3.1

sqlparse==0.5.3

Twisted==25.5.0

txaio==25.6.1

typing_extensions==4.14.1

tzdata==2025.2

uritemplate==4.2.0

urllib3==2.5.0

zope.interface==7.2

gunicorn==21.2.0

uvicorn[standard]==0.30.1

whitenoise==6.6.0

dj-database-url==2.1.0
